+++ CHANGELOG.md
@@ -27,7 +27,7 @@ The minor version will be incremented upon a breaking change and the patch versi
 - examples: add progress bar to client tool ([#456](https://github.com/rpcpool/yellowstone-grpc/pull/456))
 - proto: change error type in mod `convert_from` ([#457](https://github.com/rpcpool/yellowstone-grpc/pull/457))
 - proto: add mod `plugin` with `FilterNames` cache ([#458](https://github.com/rpcpool/yellowstone-grpc/pull/458))
-- proto: move enum Message from geyser code ([#459](https://github.com/rpcpool/yellowstone-grpc/pull/459))
+- proto: move enum Message from geyser crate ([#459](https://github.com/rpcpool/yellowstone-grpc/pull/459))
 
 ### Breaking